+++ flink-python/src/main/java/org/apache/flink/table/runtime/functions/python/PythonScalarFunctionFlatMap.java
@@ -20,42 +20,26 @@ package org.apache.flink.table.runtime.functions.python;
 
 import org.apache.flink.annotation.Internal;
 import org.apache.flink.api.common.functions.RichFlatMapFunction;
-import org.apache.flink.api.common.typeinfo.TypeInformation;
-import org.apache.flink.api.java.typeutils.RowTypeInfo;
 import org.apache.flink.configuration.Configuration;
 import org.apache.flink.python.PythonFunctionRunner;
 import org.apache.flink.table.functions.ScalarFunction;
-import org.apache.flink.table.functions.python.PythonEnv;
 import org.apache.flink.table.functions.python.PythonFunctionInfo;
 import org.apache.flink.table.runtime.runners.python.scalar.PythonScalarFunctionRunner;
 import org.apache.flink.table.types.logical.RowType;
-import org.apache.flink.table.types.utils.TypeConversions;
 import org.apache.flink.types.Row;
-import org.apache.flink.util.Preconditions;
 
 import org.apache.beam.sdk.fn.data.FnDataReceiver;
 
 import java.io.IOException;
-import java.util.Arrays;
 
 /**
  * The {@link RichFlatMapFunction} used to invoke Python {@link ScalarFunction} functions for the
  * old planner.
  */
 @Internal
-public final class PythonScalarFunctionFlatMap extends AbstractPythonStatelessFunctionFlatMap {
+public class PythonScalarFunctionFlatMap extends AbstractPythonScalarFunctionFlatMap {
 
 	private static final long serialVersionUID = 1L;
-
-	/**
-	 * The Python {@link ScalarFunction}s to be executed.
-	 */
-	private final PythonFunctionInfo[] scalarFunctions;
-
-	/**
-	 * The offset of the fields which should be forwarded.
-	 */
-	private final int[] forwardedFields;
 
 	public PythonScalarFunctionFlatMap(
 		Configuration config,
@@ -64,28 +48,7 @@ public final class PythonScalarFunctionFlatMap extends AbstractPythonStatelessFu
 		RowType outputType,
 		int[] udfInputOffsets,
 		int[] forwardedFields) {
-		super(config, inputType, outputType, udfInputOffsets);
-		this.scalarFunctions = Preconditions.checkNotNull(scalarFunctions);
-		this.forwardedFields = Preconditions.checkNotNull(forwardedFields);
-	}
-
-	@Override
-	public void open(Configuration parameters) throws Exception {
-		super.open(parameters);
-
-		RowTypeInfo forwardedInputTypeInfo = new RowTypeInfo(
-			Arrays.stream(forwardedFields)
-				.mapToObj(i -> inputType.getFields().get(i))
-				.map(RowType.RowField::getType)
-				.map(TypeConversions::fromLogicalToDataType)
-				.map(TypeConversions::fromDataTypeToLegacyInfo)
-				.toArray(TypeInformation[]::new));
-		forwardedInputSerializer = forwardedInputTypeInfo.createSerializer(getRuntimeContext().getExecutionConfig());
-	}
-
-	@Override
-	public PythonEnv getPythonEnv() {
-		return scalarFunctions[0].getPythonFunction().getPythonEnv();
+		super(config, scalarFunctions, inputType, outputType, udfInputOffsets, forwardedFields);
 	}
 
 	@Override
@@ -107,15 +70,6 @@ public final class PythonScalarFunctionFlatMap extends AbstractPythonStatelessFu
 	}
 
 	@Override
-	public void bufferInput(Row input) {
-		Row forwardedFieldsRow = Row.project(input, forwardedFields);
-		if (getRuntimeContext().getExecutionConfig().isObjectReuseEnabled()) {
-			forwardedFieldsRow = forwardedInputSerializer.copy(forwardedFieldsRow);
-		}
-		forwardedInputQueue.add(forwardedFieldsRow);
-	}
-
-	@Override
 	public void emitResults() throws IOException {
 		byte[] rawUdfResult;
 		while ((rawUdfResult = userDefinedFunctionResultQueue.poll()) != null) {
@@ -124,10 +78,5 @@ public final class PythonScalarFunctionFlatMap extends AbstractPythonStatelessFu
 			Row udfResult = userDefinedFunctionTypeSerializer.deserialize(baisWrapper);
 			this.resultCollector.collect(Row.join(input, udfResult));
 		}
-	}
-
-	@Override
-	public int getForwardedFieldsCount() {
-		return forwardedFields.length;
 	}
 }
